Initial commit for script to organize comprehensive testing of PLplot for all platforms with access to bash and GNU make.
For each of the three principal PLplot configurations, this script
runs all 7 of our fundamental tests.  Those three configurations
are (1) shared library/dynamic devices; (2) shared library/nondynamic
devices; and (3) static library/nondynamic devices.  The 7 fundamental
tests are ctest in the build tree and running the make targets
test_noninteractive and test_interactive in (1) the build tree; (2) the
installed examples tree with cmake build system; and (3) the installed
examples tree with traditional (Makefile + pkg-config) build system.

FIXME: currently the choice of which configurations and which of the 7
possible tests to run is controlled by some hard-coded variables. 
Those variables are currently set to use all all 7 tests for each of
the three configurations. Those variables should be controlled by
user-settable options for maximum flexibility.


svn path=/trunk/; revision=11140