For Steve Schwartz: Apply plarc_approx bugfix
From Steve's email:

1  When you filled an arc, you did so by OVERWRITING the last point in
the set of arc segments by the origin, so the filled arc stops short of
the last theta.

2  While you check to ensure that there are at least 2 segments so you
can fill the polygon, again since you overwrite the last point you only
end up with a polygon with 2 points which plfill can't fill.

3  When you calculate the number of segments, you don't allow for an arc
drawn backward, and so end up with a negative number of segments, which
then turns into 2 which is still not enough (see (2)).

svn path=/trunk/; revision=11192