Complete rewrite of pllegend.
The API now includes the following:

Line, symbol, and cmap0 legends.  Line and symbol are completely
implemented and tested.  cmap0 should be a complete API, but it is not
implemented yet. A cmap1 API is still in the planning stages for
pllegend.

A (semitransparent) background option. Note, the extent of the
background depends on the size of the character strings that are
rendered so more work is required on plstrl and various device drivers
before the background option gives properly aligned results for
anything but the Hershey fonts.

The text part of the legend has the following overall characteristics:
text offset (between text and plot), scale (of vertical height of
characters), (vertical) spacing, justification, and (cmap0) color
index.

Any particular legend index can have any combination of a line,
a line of symbols, and/or a cmap0 box.  The line characteristics
are (cmap0) color index, line style index, and line width.  The symbol
characteristics are the number of symbols, the (cmap0) color index, and
the scale (of vertical height of the symbol).  The cmap0 box
characteristics are height, (cmap0) color index, and fill pattern style.


svn path=/trunk/; revision=11210