Drop cmap1 color box option so that all colors are expressed with cmap0.  That is a nice simplification as discussed on list.
Improve handling of PL_LEGEND_NONE so that it overrides all other
opt_array flags.

Reorder/rename some arguments in a more rational manner.  For example,
the color box, line, and symbols for each legend entry are rendered
in that order so put the arguments in that order as well.


svn path=/trunk/; revision=11231