Convert from /* ... */ commentary style to // style for all C (as opposed to C++) headers.  The results pass a number of tests that are mentioned on list.
ToDo: So far we have only converted C code in src and include to the
new commentary style.  Those are obviously really important cases, but
we still have lots more C and C++ code (e.g., in lib, bindings,
drivers, and examples) that needs to be changed to the new commentary
style.


svn path=/trunk/; revision=11269