Ignore all instances of a string (such as "/*") being searched for on a line where the first such instance is quoted.
This rule obviously is not a completely general solution of all the
potential problems associated with quotes, but it gets the job done
for one of the lib files where this turned out to be important. Also,
a search for bad combinations of '//' and '"' in the resulting lib
results (as well as the historical src and include results) showed
no issues.


svn path=/trunk/; revision=11270