Implement plstring which is designed to supersede plsymbol and plpoin when the glyph(s) you want to plot at each of the points can be expressed with a standard user input string.  As with plmtex or plptex the standard input string can be UTF-8, or the ascii subset of UTF-8, or one of the standard PLplot text escape sequences.
N.B. the use of plstring highlights that there are some vertical
alignment issues with plptex (which is called internally by plstring)
depending on the device driver that is being used.

Change the API of pllegend to express the symbols (glyphs) used for
each legend entry as a standard user input string (which allows
pllegend to call plstring internally).

Change the 4th and 26th example to use this new pllegend API and to
also use plstring to plot the line in the non-legend part of the plot.
Also, make the legend background colour more transparent (alpha
changed from 0.90 to 0.70).


svn path=/trunk/; revision=11310