Tweak directions to set PLplot release version number of doxygen-generated documentation.
Also, tweak directions to account for the important change in
scripts/generate_website.sh which copies all files under svn control
(including all local changes to those files) from the source tree that
includes that script.  This change, means for example, that as a
matter of convenience the release manager (or anybody else) does not
have to commit local changes in order to generate and test the
website based on those local changes.


svn path=/trunk/; revision=11346