Initial commit of swig-generated bindings for Octave.
N.B. these bindings are ignored and the traditional matwrap-generated
bindings used by default unless the user specifies
-DENABLE_swig_octave=ON.

The only tests run so far with -DENABLE_swig_octave=ON on the Linux
platform are the plplot_octave target builds
bindings/octave/plplot_octave.oct and the ldd -r result on
plplot_octave.oct is identical to that for the traditional
matwrap-bindings (except for the index numbers for each library).

Further planned tests are to see if the standard octave example 10
works with these new bindings, and if so, try all the same tests on
the wine version of Windows.


svn path=/trunk/; revision=11385