Initial steps toward a plcolorbar implementation
The eventual plan is to use this function to support color bars for plimage,
plshades and possibly plgradient.

At this point, plimage is partially supported.  Support for plshades and
plgradient are next on my TODO list.

- The position of the color bar can be specified relative to any one of the
  four sides of the plot sub-page.
- The length and width of the color bar are specified by the user.
- An array with a high and low value can be passed, which will result in
  a plimage-based color bar with automatically-interpolated steps between
  the low and high values.
- A label can be drawn on a user-specified side of the color bar
- End-caps can be drawn on either end of the color bar
- Two new pages have been added to example 33, illustrating some basic
  functionality.

This requires a lot of testing, particularly the various combinations of
color bar position, label positions and inclusion/exclusion of end caps.

svn path=/trunk/; revision=11394