Improved octave configuration for the Windows case.
Find path of hdf5.h.  (We got away with not finding the path to this
header on Linux because it is in a standard location, /usr/include.)

Take more care with PATH issues so that the native form is tranformed
to the CMake internal form.  (This is a no-op on Linux, but it
translates the native \ form to CMake / form of path on Windows.)


svn path=/trunk/; revision=11397