Update plcolorbar with proper labeling and tick mark support for shade bars
The implementation could probably be cleaned up a bit.  The function
label_box_custom is a copy of label_box, modified to support user-provided
arrays of x- and y-positions for labels AND tick marks.  It is a sort of
plaxes-light, and not really meant for use outside of internal PLplot
functions at this time.

svn path=/trunk/; revision=11430