Implement proof of concept (with a test function, testppchar) of how to deal with Octave character arrays as arguments. This should prove useful for bindings of pllegend and plstripc (currently handled in a clumsy way by wrapping with a function that takes four character string arguments rather than one argument of a characer array containing four strings).
Currently the char **ArrayCk typemap only handles the charMatrix case
(an ordinary Octave matrix filled with strings that have been extended
with blanks so they all have the same length).  Additional logic to
detect the cell matrix case and deal separately with it should be
implemented as well.  Also, logic to handle other error conditions
should be implemented.  But the current typemap should be good enough
to get all the pllegend-using examples to work for now using the
non-cell character matrix case.


svn path=/trunk/; revision=11437