Initial commit of 33rd example for octave.  This runs without issues from a hand-configured octave environment (to take care of global variable issues) for -dev xwin and gives results that are indistinguishable by eye from the C -dev xwin results for this example.  However, there is some malformed UTF-8 evident for unicode-aware devices like -dev psc.  The error message is
*** PLPLOT ERROR, ABORTING OPERATION ***
UTF-8 string is malformed: � aborting operation

repeated often.  That issue obviously still needs to be sorted out.

After that dealing with global variable issues must be automated
(using the recently updated plplot_octave_def file), and test scripts
adjusted to test example 33 with octave to make sure when all
corrections are completed it is giving the same results as the
corresponding C example.


svn path=/trunk/; revision=11453