Implementation of the Fortran 95 interface to pllegend. Extended example 4 with the appropriate calls. No formal comparisons yet.
Note: the interface uses the nlegend arguments just as the C version. I could have 
used the size of the text argument to determine this, but that would have meant 
that the user has to select the right section, e.g. call pllegend(..., text(1:nlegend), ...)
to limit the number of entries. It would not have been as convenient as with the other
cases where a "size" argument is derived on the Fortran 95 side. 

svn path=/trunk/; revision=11544