Corrections in example 33 for F77, (more or less) completed work on F95 binding for pllegend() and the examples. They produce results that are identical to the C examples.
The interface for F95 needs some tweaking: we do not want the nlegend argument and we need additional
checking of the arguments. This means example 4 will change a bit and the implementation of the interface will 
change (more sanity checking). Rationale: we want to take advantage of more Fortran 95 features. 

svn path=/trunk/; revision=11547