First set of Tcl examples adjusted: add {} for expressions and remove local definitions of PLplot symbols.
Note: in example 8 I ran into an oddity with the hexadecimal notation for MAG_COLOR. I had
to use [expr] to get the picture right. I think it is due to the use of the C function atoi()
to convert a string to a number. We should replace it with Tcl_GetInt() to take care of 
the alternative forms.

svn path=/trunk/; revision=11572