Replace drawConvexPolygon with drawPolygon.  This fixes a -dev qtwidget bug for example 27 which has complex (i.e., not convex) boundaries for the filled polygons that are plotted.
Use Qt::WindingFill fill rule rather than the (default)
Qt::OddEvenFill fill rule.  This makes a major difference for the
example 27 fill results. For some reason the former is implemented in
the Qt4 library in a much less buggy way than the latter.

N.B. the Qt::WindingFill fill rule is ignored by the svgqt device.
Instead, the resulting svg file has the incorrect oddeven fill
attribute.  This must be a bug in the Qt4 library implementation (at
least for the 4.6.3 Debian squeeze version of Qt4).


svn path=/trunk/; revision=11579