Change XFillPolygon shape argument from Nonconvex to Complex since the latter is required for possibily self-intersecting boundaries (as occurs for example 27) according to the XFillPolygon man page.  In fact, the man page says the fill results are undefined for self-intersecting boundaries if you use the incorrect Nonconvex shape argument rather than the correct Complex one.
This change made no difference to the rendering errors that occur for
example 27 for either the normal WindingRule fill rule or if that is
changed to the EvenOddRule fill rule.  Nevertheless, I am making this
change to be in conformance with the requirement as set out in the man
page and just in case it does make a difference for X server platforms
other than my Debian Squeeze one for Intel video hardware.


svn path=/trunk/; revision=11581