Fix up some more of the PL_MAXPOLY logic. This change allows fills for boundaries defined by large numbers of points (e.g., example 27) to run correctly (i.e., without hanging) for -dev tk.
N.B. There is still more PL_MAXPOLY work to do for the plr_line
routine in plt.c.  The issue there is the code is so complicated it is
difficult (for me, at least) to figure out how much memory needs to be
allocated.  Fortunately, this accident waiting to happen does not seem
to be triggered by our current examples.


svn path=/trunk/; revision=11582