Update C and python versions of example 27 so that fills work correctly both for -eofill (even-odd fill) and with that option dropped (non-zero winding rule fill).
All the previous fill issues for self-intersecting boundaries were an
artifact of the incorrect angle termination condition (via windings).
The correct windings must be calculated such that _both_ phi and phiw
are the minimum possible integer multiple of 2 pi.  It is easy to show
that this condition leads to windings = params[1]/gcd(params[0],
params[1]) (with appropriate adjustments to convert to integers and
use absolute values).

With this change to the maximum phi and phiw, the results look
beautiful for both -eofill and not.  

I ascribe the previous bad filling results for the -eofill case as due
to the boundaries of certain areas being duplicated an even number of
times (which reverses the decision about whether a given area will be
filled or not for the even-odd fill rule).  I ascribe the previous
slightly bad fill results for the non-zero winding number fill rule
to the last large boundary segment that was required to close the
boundary for the previous case.

A side benefit of no longer using an excessive number of windings is
that NPTS can be divided by 10 and yet still have a smooth result.

I also changed the xmin, xmax, ymin, ymax logic so that it worked
reliably for all cases.

For the python case I also took advantage of numpy array capabilities.

The resulting C and python results are consistent.  Please propagate
these windings, NPT, and min/max changes to all other languages.



svn path=/trunk/; revision=11594