Complete the change to use standard code delimiters "%{" and "%}" inside %define ... %enddef swig directives.  Also, correct what appears to be a parentheses error.  That unbalanced parentheses issue also appears in /usr/share/swig1.3/java/typemaps.i for the
%typemap(directorin,descriptor=JNIDESC) TYPE &OUTPUT

typemap, but not for other related typemaps so I may be misunderstanding
something.  But when I provide balanced parentheses as for the
other cases, the result continues to work (see below).

Also change to form of warning message in
/usr/share/swig1.3/java/typemaps.i for the 

%typemap(directorin .....) TYPE *OUTPUT

typemap.

These changes appear to make no difference to build and test results,
but there is a good possibility they will improve uncrustify styling
results.


svn path=/trunk/; revision=11699