Put the orientation of the colorbar directly under the control of the user with the PL_COLORBAR_ORIENT_(RIGHT, TOP, LEFT, BOTTOM) orientation bits of opt.  However, if the user does not specify any of these bits, then adopt a sensible orientation default which depends on position.
Make all previous orientation logic (which was controlled strictly by
position bits before) depend on the orientation bits instead.

There are no changes to the example 33 PostScript results due to these 
internal changes (because of the default orientation choice noted
above).

ToDo: My goal here is that position should control nothing except
overall position of colorbar with no cross-talk with how that colorbar
object is specified except for the default value of the orientation. 
I have only dealt with the obvious orientation stuff for now and there
is some non-position logic (e.g., label offsets, numerical label
positions, and tick marks for the colorbar) that currently depends on
position bits that still must be dealt with to achieve my goal.


svn path=/trunk/; revision=11736