If both "b" and "u" options specified for xopt or yopt argument for plaxes (or plbox), "u" option has the priority.
If both "c" and "w" options specified for xopt or yopt argument for
plaxes (or plbox), "w" option has the priority.

Implement "x" option for xopt or yopt argument for plaxes (or plbox).
This option has all the side effects of "t" such as numerical tick
labelling (if specified), but the actual ticks are not drawn.


svn path=/trunk/; revision=11747