Add drawing mode setting and getting support (no driver support yet)
These two new functions allow a user to specify different drawing modes,
such as replacing the current plot content with the drawing color, even if
the drawing color is transparent (SOURCE in Cairo) or XOR drawing.

This is only intended to be supported by output devices which support these
operations directly.  At least Cairo and Qt support the modes included in
this commit, along with several others.

svn path=/trunk/; revision=11751