Add drawing mode setting/getting support for Cairo output devices
There are a few changes in how transparent colors are handled by the xcairo
device in order to keep output consistent with the other Cairo devices.
This should not affect any of the existing examples, and it may fix/avoid
some potential rendering bugs which could come up with the xcairo device.

svn path=/trunk/; revision=11752