Style changes:
Remove whitespace just inside cast parentheses.  This is the style we
already have in most cases, but force this in all cases now for
consistency.

Force removal of any spaces between pointer stars.  (i.e., "* *" ==>
"**").  This is our uniform style now in any case (i.e. it introduces
no source code changes).  However, this change is needed for when (in
the immediate future) we move to uncrustify-0.58 to work around an
issue (all "**" instances were being changed to "* *" inside casts)
for that version of uncrustify.


svn path=/trunk/; revision=11759