API change for plcolorbar that adds functionality that is already available in pllegend.  New arguments to return bounding box size of colorbar, specify a background color for the area contained by the bounding box (if the PL_COLORBAR_BACKGROUND bit is set in opt), and specify the line color and line style of a boundary line for the bounding box (if the PL_COLORBAR_BOUNDING_BOX bit is set in opt).
Add internal infrastructure to support complete position functionality
that is the same (except for defaults) as already implemented in
pllegend.  

ToDo: the new position functionality is a work in progress with some
remaining coordinate transformation issues so, for example, the size
of the bounding box is currently only correct in the simplest cases,
the bounding box is not yet plotted in the correct place, and the caps
are also not yet plotted in the correct place with the correct size. 
But the 16th and 33rd examples do build without errors and run (using
the -colorbar option) without showstopper errors such as segfaults.


svn path=/trunk/; revision=11763