Fix minor bug in pllegend routine where line style changes could be propagated outside the routine.
Drop old version of plcolorbar code since the examples have long since
been inconsistent with that version.

Improve specification of cap proportions using cap_angle, the angle in
degrees of the angle of the cap triangle that points in the direction
of the orientation of the cap.

Correctly position boundary box and caps according to the new
positioning scheme.  This last is an interim commit to preserve my
current work, and there are a lot more changes to come to correctly
position bar, tickmarks and numerical labels for that bar, and the
descriptive text. Also, the boundary box currently just refers to the
inner bar without extensions (yet) due to caps, numerical labels, and
the descriptive text.

The result builds without issues, and examples 16 and 33
run with with no run-time errors for the case where
the -colorbar option is used.


svn path=/trunk/; revision=11764