Update colorbar, tickmarks and numerical labels, and text to use the new positioning scheme for plcolorbar.
ToDo: there is still work required to understand and fix the
positioning of the shades case of colorbar and to expand the bounding
box to account for more than just the core colorbar.

However, if you ignore the shades case (i.e., focus on just the image
and gradient cases) for example 33, the results look reasonable again
so I am getting there.


svn path=/trunk/; revision=11766