Implement infrastructure for bounding-box calculation for plbox, label_box, and label_box_custom (including decorations such as inverted tick marks and numerical tick labels).  Use this infrastructure to help calculate the position of the complete colorbar (including decorations) and its bounding box.  For now, the revised example 33 shows the effects of inverted tick marks on position and bounding box of the complete colorbar.
ToDo: for now, I have only implemented and tested the inverted tick
mark case as a proof of concept for the plbox bounding-box
infrastructure.  Thus, additional work needs to be done to include the
numerical tick labels in the bounding box for plbox.  Also, the
colorbar label (if present) is independent of plbox so must be
included independently into the bounding box calculation for the
complete colorbar.


svn path=/trunk/; revision=11774