I have discovered experimentally that contrary to the documentation at http://developer.gnome.org/pango/unstable/PangoMarkupFormat.html, rise is in units of 1024th of a point (i.e., a real distance unit) rather than in units of the (variable) font size.  Implement a correction the accounts properly for the real units of rise. The result is the actual superscript/subscript vertical offsets that are rendered by cairo devices now is (correctly) proportional to the font size (as opposed to being fixed in size before).  This correction solves the second issue noted in the previous commit.
The cairo and PLplot character coordinate system have vertical zero
points of respectively the baseline of the fonts and the middle of the
vertical dimension of the fonts. This mismatch in coordinate systems
means the superscript/subscript offsets appear asymmetrical for cairo
devices compared to -dev xwin (i.e., issue 1 noted in the previous
commit message).  An approximation for this coordinate shift has been
implemented, and the asymmetry appears to be greatly reduced as a
result.


svn path=/trunk/; revision=11781