Implement a private plP_script_scale function that calculates scale values for font size and magnitude of the vertical offset associated with superscripts/subscripts.  Currently, the font scale and offset calculations associated with superscripts and subscripts are done in an idiosyncratic (and often incorrect) way in various locations within the core code and drivers.  plP_script_scale is designed to consolidate those bits of logic in one place where the calculation is done correctly.
As a first use of this new logic, use plP_script_scale for the Hershey
font superscripts/subscripts. plP_script_scale produces completely
symmetrical vertical offsets between superscripts and subscripts.
Therefore, this change corrects a very long-standing error (since
before 1993) in the vertical offset of the Hershey subscripts which
previously were not consistent with the magnitude of the vertical
offsets of the superscripts. 


svn path=/trunk/; revision=11783