Use plP_script_scale method to replace idiosyncratic method of determining superscript/subscript font sizes and offsets.  The test_superscript_subscript.py results are much improved by this change.
However, a vertical offset issue for the pdf device driver that was
evident before this change still continues; superscript offsets shift
the entire string downwards (including all text before when "#u" is
used) while the corresponding result for subscripts (when "#d" is
used) has no such effect.  I have looked hard for the origin of this
pdf bug but cannot find the source of it.  The superscript and
subscript paths are absolutely identical in the code except for the
numerical calculation of dev->yOffset.  So why does the correct offset
from the baseline being calculated in both cases affect the overall
baseline (including the unsuperscripted/unsubscripted _prior_ parts of
the string) in the superscript case but not in the subscript case?


svn path=/trunk/; revision=11804