Fix accident waiting to happen for get_subpage_per_mm (a static function which is called directly or indirectly by pllegend and plcolorbar).
The only reason this function worked correctly before was it was always
called after

plvpor( 0., 1., 0., 1. );
plwind( 0., 1., 0., 1. );

But the point is the get_subpage_per_mm doesn't need to know anything
about the viewport, and a call to plgspa returns all data required to
calculate the x and y ratios of normalized subpage coordinates to mm.
So get_subpage_per_mm can be substantially simplified, and that
simplification insures that routine will give correct results when
plvpor and plwind have been called with other than the above values.


svn path=/trunk/; revision=11813