Code reorganization in preparation for calculating complete bounding box and position of the labelled and decorated colorbar.
Systematically use the mm form of bounding box limits,
plsc->boxbb_xmin, etc., rather than the confusing mixture of that form
and one where the limits are expressed in normalized subpage
coordinates.

Implement the static function, calculate_limits, to isolate in one
place all the limit changes due to a changed bounding box.  The first
call of this function is associated with changing the bounding box and
corresponding position calculation from the undecorated colorbar to
the decorated (ticks, numerical tick labels, and exponents) colorbar.
The second call of this function (not yet used) will be associated
with changing the bounding box and corresponding position
calculationfrom the decorated colorbar to the labelled and decorated
colorbar.

However, the calculate_limits code is not yet fully
understood/documented (since it just collected pre-existing code which
was not fully understood/documented into a static function) so the
second call is not used yet.

As expected, the result is identical to the previous commit, i.e.,
labels are currently ignored in the bounding box calculation and in
the calculation of the position of that bounding box in the overall
plot.


svn path=/trunk/; revision=11817