Improve internal documentation of calculate_limits so that it is more obvious what is being calculated.
Distinguish arguments for calls to calculate_limits more carefully
to keep track of everything necessary for calculating offsets
and sizes of bounding boxes for undecorated colorbar, decorated
colorbar, and labelled and decorated colorbar.

Implement bounding box calculation in draw_label so that the overall
bounding box (for labelled and decorated colorbar) will (finally) be
correct.  This change has been tested for long labels for plcolorbar
in example 33 where both x and y bounding box limits are increased as
well as the standard short labels for plcolorbar in example 33.

This commit completes my planned work on the bounding box and
position calculations for plcolorbar.


svn path=/trunk/; revision=11820