Experimental support for building OCaml bindings with a static PLplot core
The OCaml bindings could only previously be compiled against a shared
library version of PLplot.  This commit updates the CMake rules so that, at
least on my system (Ubunty 11.04, 64bit), the OCaml bindings and examples
build and run properly when -DBUILD_SHARED_LIBS=OFF is passed to cmake.

Testing on other distributions and OSs is welcome!

svn path=/trunk/; revision=11884