Fix for Windows broken build for the 5.9.8 release.  This fix supplies a needed lt_dlmakeresident routine for ltdl_win32.c so that dynamic devices can be built on Windows.  Currently, this version of lt_dlmakeresident does nothing but return success.  The net result is that on Windows plend will unload all libraries that our dynamic devices depend on just like what happened for 5.9.7.
Note that for Linux we found that if we did not call the
lt_dlmakeresident version for libltdl, that calls to plend (which in
turn call lt_dlexit()) unloaded the external libraries that are
depended on by our dynamic devices.  That library unloading sometimes
lead to exit handler errors for those libraries on Linux.  

In the future if library unloading also leads to issues for Windows,
we will need to add real functionality to the lt_dlmakeresident
function in ltdl_win32.c to also avoid unloading the libraries in that
case.  However, we have no reports of such library unloading
issues on Windows so a lt_dlmakeresident version for ltdl_win32.h that
does nothing other than return success is all we appear to need at
the present time.


svn path=/trunk/; revision=11938