Completely reorganize the script.
Now exits immediately if there is a detected error condition rather
than using a bunch of Boolean logic with appropriate massive
indentation to implement the various exits.

Now more care is used with the order of the commands executed and what
is cleaned to reduce the high-water mark of disk usage as much as
possible.  These changes mean, e.g., the interactive tests will
require rebuilds of things that were cleaned previously, but that is a
relatively small sacrifice in cpu time to keep the high-water mark
(currently ~3-4GB) smaller than the alternative without these changes.


svn path=/trunk/; revision=11963