Bump minimum version of CMake to 2.8.2 for Linux and 2.8.6 for all other platforms as discussed on the plplot-devel mailing list.
I have just confirmed that CMake-2.8.6 gives good results on Linux
(using the ctest test), but that also needs to be tested on our other
platforms.  As soon as that platform testing is complete, I plan to
remove the CMake-2.6.x logic and files that are currently part of our
build system.


svn path=/trunk/; revision=12000