PC_REQUIRES_TAG always set locally so drop setting a default global value for that variable/
Improve commentary concerning the NON_TRANSITIVE option.

Implement the FORCE_EXTERNAL_STATIC option for the (internal) static
library case (NOT BUILD_SHARED_LIBS) to force CMake to find and use
only external libraries with a *.a suffix (i.e., static libraries).
This suffix trick only works with UNIX so FORCE_EXTERNAL_STATIC is
ignored in all other cases. The option is experimental.  In fact, it
still does not completely work at this time because of pkg-config
complications. Therefore, its default value is OFF.


svn path=/trunk/; revision=12055