Modify cmake_link_flags module to signal success/failure with the CMake logical truth of the link flags variable that is returned rather than using "message(FATAL_ERROR ..." to stop all cmake processing on failure.
Modify pkg_check_pkgconfig macro to honor FORCE_EXTERNAL_STATIC
variable.  If that variable is true, then the static versions of all
variables are used and checked, e.g., with a call to cmake_link_flags.
Also, signal the success/failure of pkg_check_pkgconfig with the CMake
logical truth of the link flags variable that is returned.


svn path=/trunk/; revision=12057