Provide NON_TRANSITIVE ON option only for the case of shared libraries.
For the internal static library case, transitive linking is used
regardless by CMake so the only effect of NON_TRANSITIVE ON for this
case would be non-transitive linking for pkg-config used by the
traditional build and test of the installed examples.  Such linking
can introduce inconsistencies with the linking and software library
detection of the core PLplot build.  So avoid that potential
inconsistency by simply not allowing NON_TRANSITIVE ON for the
static internal library build for PLplot.


svn path=/trunk/; revision=12066