Update build system so that the plf95demolib library works properly in the build tree, the installed examples built and tested with our CMake-based build system and the installed examples built and tested with the traditional Makefile + pkg-config approach.
For now I have chosen the approach of installing the library and module in
the standard places, but it might be better to install those
in $prefix/share/plplot5.?.?/examples/f95 since they are designed to
be completely specific with regard to our Fortran 95 examples.

I have also chosen to link to the new plf95demolib library outside the
pkg-config system for our traditional build of the installed f95
examples by configuring the Makefile appropriately.



svn path=/trunk/; revision=12071