For the --apply option change the shell script logic to do the following steps:
1. List all files (and diffs all files if the --diff option is used as
well) that will be changed.

2. Ask whether you want to apply those changes.  If no, immediate exit.
If yes, proceed with the further steps.

3. Apply the changes.

4. List all files (and diffs all files if the --diff option is used as
well) that required further styling after the current style changes
were made.  Normally, this step does not list anything, but it does
warn you when further runs of the script are required on those rare
occasions where uncrustify has to be applied iteratively.


svn path=/trunk/; revision=12127