Replace the iargc and getarg GNU extensions in bindings/f95/configurable.f90 with the Fortran 2003 standard library routines, command_argument_count and get_command_argument.
Arjen's tests show this change means it is no longer necessary to
build configurable.f90 in a separate static library on windows
platforms.  Therefore, remove all STATIC_OPTS logic from the build
system that made that separate static build possible. This means,
configurable.f90 is now completely equivalent to other source for the
plplotf95 library on all platforms now.  (Since it is not configured,
it is also time to change its name, but I will leave that to Arjen.)

This change has been tested with the test_diff_psc target on Linux
with good results (except for examples 20 and 28 because of on-going
blank issues and the missing example 00).  That target should also be tested
on Windows to make sure all is well with that platform for Fortran 95.


svn path=/trunk/; revision=12160