Add the constants that are defined via DATA statements to the list of exported symbols. This is required by the MS VC/C++ - Intel Fortran combination, but not by gcc/gfortran. Without these extra exports you get unresolved external symbols.
Note: the use of the hexadecimal numbers prevents us from using parameters. According to 
the Fortran 90/95 standards, you can not use such numbers for parameters. This has been
relaxed for Fortran 2003, so we may want to revisit this issue in the near future 
(both gfortran 4.6 and later and Intel Fortran 11.1 and later support such numbers in
the parameter statement.) 

svn path=/trunk/; revision=12171