Update branch to be consistent with tested version that works with the svn trunk version of PLplot.
Change spelling of the CMAKE_<LANG>_LINKER_PREFERENCE string to the
correct form in English ("Preferred").  This introduces no changes in
results (which is also confirmed by the cmake (2.6.0 and 2.8.1rc3)
source code where it is clear only the first letter of the
CMAKE_<LANG>_LINKER_PREFERENCE variable is relevant.


svn path=/branches/test_cmake/; revision=12229