Modify the ndiff test script that is created by CMake to include extra filtering in the Windows case to get rid of cr part of line endings and remove redundant zeros from exponents to make results more comparable to the Unix test files.
Test results on MinGW/MSYS/Wine failed in 12 or the 26 cases so instead
of using cmp -s for the tests I used diff to see what those
differences were from the Unix files.  In all cases, the results
were due to minor differences of floating-point noise in results from
ndiff on this platform and on Linux.  In virtually all cases when
the floating-point noise was different between the two platforms, the
Wine platform's noise was slightly larger.  

My working hypothesis to explain those results is that Wine is being
slighty sloppy in how well it retains floating-point precision in
calculations compared to Linux.  A similar test on Microsoft Windows
might confirm or deny that hypothesis.


svn path=/trunk/; revision=12265