Initial commit of a useful bash script to do detailed comparisons of results from two different file:// svn repos.
This script has already confirmed identical latest results, identical
log files, and identical results for a substantial sample of
historical revisions for the pre_Allura and Allura versions of the
lasi (libLASi) repos.  It will also be used for a number of other
projects culminating in checking the PLplot svn repos once they
are converted to Allura.


svn path=/trunk/; revision=12296