Disable all bindings but ocaml by default to eliminate build breakage caused by recent C API breakage when the user chooses the default bindings.
To enable these bindings again by default for their language of
choice, the interested developer should propagate the recent integer
==> double line width changes in the C API for pl*shade*, pllegend,
and plcolorbar to their bindings of choice, make the appropriate
changes in examples 4, 16, 26, and 33 to use the changed API, and test
the result (at least by running the test_noninteractive target) with
the language specifically enabled, e.g., with the -DENABLE_python=ON
option.  If that testing works, then the final step is to enable the
language by default again, and commit your changes.



svn path=/trunk/; revision=12314