Initial version of "build projects" project which configures a build of shapelib with no issues on either Linux or MinGW/MSYS/Wine and which also includes an initial version of the build configuration for ndiff which needs to be updated to be consistent with the style used for shapelib.
These files currently configured to refer to local tarballs or zip
files on my system to avoid exessive downloads while debugging these
configurations, but these should be changed to the download versions
as soon as these configurations mature a bit to the point where others
will want to use them.

Forgot to svn add and commit this essential file for this initial
version.


svn path=/trunk/; revision=12352